Update ComboboxMulti. Clean up code
